+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - [Feature] Reset tab on double tap
 - [Feature] Add back button in options screen
 - [Feature] Hide bottombar when keyboard close
+- [Feature] New report bug screen
 - [GitHub] Update github repo picture
 - [CI] Migrate to Detekt
 - [CI] Android lint in CI